feat: 草稿优化自动删除逻辑

+++ pamirs-core/pamirs-draft/pamirs-draft-core/src/main/java/pro/shushi/pamirs/draft/core/api/DefaultDraftApi.java
@@ -1,7 +1,9 @@
 package pro.shushi.pamirs.draft.core.api;
 
+import com.alibaba.fastjson.TypeReference;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
+import pro.shushi.pamirs.draft.api.utils.DraftDeleteFunction;
 import pro.shushi.pamirs.framework.connectors.data.sql.Pops;
 import pro.shushi.pamirs.meta.api.Models;
 import pro.shushi.pamirs.meta.api.core.orm.systems.DraftApi;
@@ -13,6 +15,12 @@ import pro.shushi.pamirs.meta.base.BaseModel;
 import pro.shushi.pamirs.meta.common.spi.Spider;
 import pro.shushi.pamirs.meta.enmu.FunctionOpenEnum;
 import pro.shushi.pamirs.meta.enmu.FunctionTypeEnum;
+import pro.shushi.pamirs.meta.util.JsonUtils;
+
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * @author Gesi at 14:55 on 2025/9/17
@@ -71,6 +79,15 @@ public class DefaultDraftApi implements DraftApi {
         if (dbDraft == null) {
             return null;
         }
+        String callDeleteFunctionName =
+                Optional.ofNullable(DraftDeleteFunction.get()).map(pro.shushi.pamirs.meta.api.dto.fun.Function::getName).orElse(null);
+
+        // 删除方法调用且找不到主键不删除草稿
+        if (StringUtils.isNotBlank(callDeleteFunctionName) && callDeleteFunctionName.startsWith("delete") && JsonUtils.parseObject(dbDraft.getDataPks(), new TypeReference<List<Object>>() {
+        }).stream().noneMatch(Objects::nonNull)) {
+            return null;
+        }
+
         return Boolean.TRUE.equals(dbDraft.deleteById()) ? data : null;
     }
 
@@ -81,7 +98,8 @@ public class DefaultDraftApi implements DraftApi {
             // 先尝试通过模型获取spi
             try {
                 draftContextApi = Spider.getExtension(DraftContextApi.class, model);
-            } catch (Exception ignored) {}
+            } catch (Exception ignored) {
+            }
         }
         if (draftContextApi == null) {
             draftContextApi = Spider.getExtension(DraftContextApi.class, "defaultDraftContextApi");
